+++ ms-auth/src/main/java/cl/duoc/cloud/auth/dto/LoginDTO.java
@@ -1,0 +1,4 @@
+package cl.duoc.cloud.auth.dto;
+
+public record LoginDTO(String username, String password) {
+}