+++ ms-auth/src/main/java/cl/duoc/cloud/auth/model/Usuario.java
@@ -1,0 +1,12 @@
+package cl.duoc.cloud.auth.model;
+
+import java.util.List;
+
+public record Usuario(
+        String username,
+        String hashPassword,
+        String nombre,
+        String email,
+        List<String> roles,
+        List<String> scopes) {
+}